Support linking against a binary distribution of LLVM (and other stuff)
Now we don’t need to build the whole of LLVM and Clang together with
our code.
This means I can build the thing on my poopy 4-GB machine.
Many thanks to Eli Bendersky for the makefile that strongly inspired
the makefile I wrote for this.
Many other changes are also included in this commit. I couldn’t commit
them earlier due to a mixture of technical and laziness-related reasons
− and now am too lazy to break them into separate commits.