Change definition of _possible accumulator_
To match the implementation. I did not implement my original vision,
which involved the concept of a _base_ identifier of an l-value.
Instead, I simply check whether the l-value involves any identifier
declared outside of the loop. That is an approximation of what we
really want, and we might need to implement the more precise check later.