Analyze only files whose names end in “.c”
Previously, when we were invoked on a directory, we searched for
and analyzed all contained files whose names had “.c” as a
substring. Consequently, we were catching files ending in ".cg"
and ".commons" and "README.carefully" and a bunch of other weird
stuff. So now we analyze only files whose names really do end in
“.c”.

(Actually, we also analyze files ending in “.cpp”, because you
might try analyzing C++ programs if you’re feeling adventurous.
But that’s not supported, and you probably shouldn’t do that.)