Add an empty compilation database to test-examples
Revert the change made by the previous commit, and deal with that issue in a
different way. Instead of appending `--` to the command line that invokes
reduction-detector in test.h, use an empty compilation database. Also briefly
explain that empty compilation database in the accompanying readme.

Also remove compile_commands.json from .gitignore, since we do actually want to
track that file in Git.